Data.Constraint.Char

Description

Utilities for working with KnownChar constraints.

This module is only available on GHC 9.2 or later.

Documentation

type family CharToNat (a :: Char) :: Natural where ...

Convert a character to its Unicode code point (cf. ord)

Since: base-4.16.0.0

type family NatToChar (a :: Natural) :: Char where ...

Convert a Unicode code point to a character (cf. chr)

Since: base-4.16.0.0

charToNat :: forall c. KnownChar c :- KnownNat (CharToNat c)

natToChar :: forall n. (n <= 1114111, KnownNat n) :- KnownChar (NatToChar n)