Data.Constraint.Deferrable

Description

The idea for this trick comes from Dimitrios Vytiniotis.

Documentation

newtype UnsatisfiedConstraint

Constructors

UnsatisfiedConstraint String

class Deferrable p where

Allow an attempt at resolution of a constraint at a later time

Methods

deferEither :: (p => r) -> Either String r

Resolve a Deferrable constraint with observable failure.

defer :: forall p r. Deferrable p => (p => r) -> r

deferred :: forall p. Deferrable p :- p

data (a :: k1) :~~: (b :: k2) where infix 4

Kind heterogeneous propositional equality. Like :~:, a :~~: b is inhabited by a terminating value if and only if a is the same type as b.

Since: base-4.10.0.0

Constructors

HRefl :: forall {k1} (a :: k1). a :~~: a

data (a :: k) :~: (b :: k) where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: base-4.7.0.0

Constructors

Refl :: forall {k} (a :: k). a :~: a