Data.Constraint.Deferrable

Description

The idea for this trick comes from Dimitrios Vytiniotis.

Documentation

data UnsatisfiedConstraint

Constructors

UnsatisfiedConstraint String

class Deferrable p where

Allow an attempt at resolution of a constraint at a later time

Minimal complete definition

deferEither

Methods

deferEither :: proxy p -> (p => r) -> Either String r

Resolve a Deferrable constraint with observable failure.

defer :: forall p r proxy. Deferrable p => proxy p -> (p => r) -> r

Defer a constraint for later resolution in a context where we want to upgrade failure into an error

deferred :: forall p. Deferrable p :- p

defer_ :: forall p r. Deferrable p => (p => r) -> r

A version of defer that uses visible type application in place of a Proxy.

Only available on GHC 8.0 or later.

deferEither_ :: forall p r. Deferrable p => (p => r) -> Either String r

A version of deferEither that uses visible type application in place of a Proxy.

Only available on GHC 8.0 or later.

data a :~~: b where

Kind heterogeneous propositional equality. Like '(:~:)', a :~~: b is inhabited by a terminating value if and only if a is the same type as b.

Only available on GHC 8.0 or later.

Constructors

HRefl :: a :~~: a

data (k :~: a) b :: forall k. k -> k -> * where infix 4

Propositional equality. If a :~: b is inhabited by some terminating value, then the type a is the same as the type b. To use this equality in practice, pattern-match on the a :~: b to get out the Refl constructor; in the body of the pattern-match, the compiler knows that a ~ b.

Since: 4.7.0.0

Constructors

Refl :: (:~:) k a a