Data.Constraint.Deferrable

Description

The idea for this trick comes from Dimitrios Vytiniotis.

Documentation

data UnsatisfiedConstraint

Constructors

UnsatisfiedConstraint String

class Deferrable p where

Allow an attempt at resolution of a constraint at a later time

Methods

deferEither :: proxy p -> (p => r) -> Either String r

Resolve a Deferrable constraint with observable failure.

defer :: forall proxy p r. Deferrable p => proxy p -> (p => r) -> r

Defer a constraint for later resolution in a context where we want to upgrade failure into an error

deferred :: forall p. Deferrable p :- p