Data.Constraint.Forall

Description

This module uses a trick to provide quantification over constraints.

Documentation

type family Forall p :: Constraint

A representation of the quantified constraint forall a. p a.

inst :: forall p a. Forall p :- p a

Instantiate a quantified Forall p constraint at type a.

type family ForallF p f :: Constraint

A representation of the quantified constraint forall a. p (f a).

instF :: forall p f a. ForallF p f :- p (f a)

Instantiate a quantified ForallF p f constraint at type a.

type Forall1 p = Forall p

inst1 :: forall p f. Forall p :- p f

Instantiate a quantified constraint on kind * -> *. This is now redundant since inst became polykinded.

type family ForallT p t :: Constraint

A representation of the quantified constraint forall f a. p (t f a).

instT :: forall p t f a. ForallT p t :- p (t f a)

Instantiate a quantified ForallT p t constraint at types f and a.